Updated the update user function

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get goals
 // @route GET /api/goals
@@ -37,6 +38,20 @@ const updateGoals = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id)
+
+  // Check if user exists
+  if(!user){
+    res.status(401)
+    throw new Error('User does not exist')
+  }
+
+  // Match logged in user with the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -52,6 +67,20 @@ const deleteGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Check if user exists
+  if(!user){
+    res.status(401)
+    throw new Error('User does not exist')
+  }
+
+  // Match logged in user with the goal user
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await goal.deleteOne();
